# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PValle API",
        "description": "Documentación de la API de PValle (Buses Mi Valle)",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/clientes": {
            "get": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Listar clientes",
                "description": "Display a listing of the resource.",
                "operationId": "3ab1bd0bc1c6d13508917ec1f71f52f5",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Buscar por nombre, apellido, cédula o email",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número de página",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado de clientes obtenido correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Listado de clientes obtenido correctamente."
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Cliente"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 15
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 25
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Crear un nuevo cliente",
                "description": "Store a newly created resource in storage.",
                "operationId": "22be7ccdd26efb899a619bd891090f4f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "apellido",
                                    "cedula",
                                    "email",
                                    "contrasena",
                                    "contrasena_confirmation"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "María"
                                    },
                                    "apellido": {
                                        "type": "string",
                                        "example": "Gómez"
                                    },
                                    "cedula": {
                                        "type": "string",
                                        "example": "9876543"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "maria@correo.com"
                                    },
                                    "contrasena": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secreto123"
                                    },
                                    "contrasena_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secreto123"
                                    },
                                    "telefono": {
                                        "type": "string",
                                        "example": "70011122",
                                        "nullable": true
                                    },
                                    "direccion": {
                                        "type": "string",
                                        "example": "Av. Siempre Viva 123",
                                        "nullable": true
                                    },
                                    "fecha_nacimiento": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1995-05-20",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Cliente creado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Cliente creado correctamente."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Cliente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Cliente": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nombre": {
                        "type": "string",
                        "example": "María"
                    },
                    "apellido": {
                        "type": "string",
                        "example": "Gómez"
                    },
                    "nombre_completo": {
                        "type": "string",
                        "example": "María Gómez"
                    },
                    "cedula": {
                        "type": "string",
                        "example": "9876543"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "maria@correo.com"
                    },
                    "telefono": {
                        "type": "string",
                        "example": "70011122",
                        "nullable": true
                    },
                    "direccion": {
                        "type": "string",
                        "example": "Av. Siempre Viva 123",
                        "nullable": true
                    },
                    "fecha_nacimiento": {
                        "type": "string",
                        "format": "date",
                        "example": "1995-05-20",
                        "nullable": true
                    },
                    "estado_base": {
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}